fix(modeldata): validate the opened catalog descriptor and keep FIFO test unix-only
